+++ lib/models/individual_bar.dart
