+++ lib/models/info_reader.dart
@@ -2,7 +2,7 @@ import 'dart:convert';
 import 'dart:io';
 
 import 'package:path_provider/path_provider.dart';
-import 'package:smart_home/presenter/setting_presenter.dart';
+import 'package:smart_home/presenters/setting_presenter.dart';
 
 class InfoReader{
   Future<String> get _localPath async {
