+++ lib/views/bar_graph.dart
@@ -1,6 +1,6 @@
 import 'package:fl_chart/fl_chart.dart';
 import 'package:flutter/material.dart';
-import 'package:smart_home/model/bar_data.dart';
+import 'package:smart_home/models/bar_data.dart';
 
 class MyBarGraph extends StatelessWidget {
   final List weeklySummary;
