+++ lib/views/build_user_info.dart
@@ -1,12 +1,12 @@
 import 'dart:typed_data';
 
 import 'package:flutter/material.dart';
-import 'package:smart_home/model/constants.dart';
-import 'package:smart_home/model/user.dart';
-import 'package:smart_home/presenter/language_presenter.dart';
-import 'package:smart_home/presenter/user_presenter.dart';
-import 'package:smart_home/view/custom_button.dart';
-import 'package:smart_home/view/show_diaglog.dart';
+import 'package:smart_home/models/constants.dart';
+import 'package:smart_home/models/user.dart';
+import 'package:smart_home/presenters/language_presenter.dart';
+import 'package:smart_home/presenters/user_presenter.dart';
+import 'package:smart_home/views/custom_button.dart';
+import 'package:smart_home/views/show_diaglog.dart';
 
 class BuildUserInfo extends StatefulWidget {
   const BuildUserInfo(
