+++ lib/views/index_screen.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
-import 'package:smart_home/presenter/language_presenter.dart';
-import 'package:smart_home/view/bottom_nav.dart';
-import 'package:smart_home/view/scheduling_screen.dart';
-import 'package:smart_home/view/setting_screen.dart';
-import 'package:smart_home/view/statistical_screen.dart';
+import 'package:smart_home/presenters/language_presenter.dart';
+import 'package:smart_home/views/bottom_nav.dart';
+import 'package:smart_home/views/scheduling_screen.dart';
+import 'package:smart_home/views/setting_screen.dart';
+import 'package:smart_home/views/statistical_screen.dart';
 
 class IndexScreen extends StatefulWidget {
   const IndexScreen({super.key, required this.reloadThemeMode});
