+++ lib/views/setting_screen.dart
@@ -1,10 +1,10 @@
 import 'package:flutter/material.dart';
-import 'package:smart_home/presenter/language_presenter.dart';
-import 'package:smart_home/presenter/setting_presenter.dart';
-import 'package:smart_home/presenter/user_presenter.dart';
-import 'package:smart_home/view/accounts_management_screen.dart';
-import 'package:smart_home/view/custom_button.dart';
-import 'package:smart_home/view/build_user_info.dart';
+import 'package:smart_home/presenters/language_presenter.dart';
+import 'package:smart_home/presenters/setting_presenter.dart';
+import 'package:smart_home/presenters/user_presenter.dart';
+import 'package:smart_home/views/accounts_management_screen.dart';
+import 'package:smart_home/views/custom_button.dart';
+import 'package:smart_home/views/build_user_info.dart';
 
 class SettingScreen extends StatefulWidget {
   const SettingScreen({super.key, required this.reloadThemeMode});
